+++ protzilla/constants/colors.py
@@ -1,0 +1,8 @@
+PROTZILLA_DISCRETE_COLOR_SEQUENCE = [
+    "#4A536A",
+    "#87A8B9",
+    "#CE5A5A",
+    "#8E3325",
+    "#E2A46D",
+]
+PROTZILLA_DISCRETE_COLOR_OUTLIER_SEQUENCE = ["#4A536A", "#CE5A5A"]
