--- protzilla/constants/constants.py
@@ -1,21 +1,0 @@
-from pathlib import Path
-from string import ascii_letters
-from random import choices
-
-PATH_TO_PROJECT = Path(__file__).resolve().parent.parent.parent
-PATH_TO_RUNS = Path(f"{PATH_TO_PROJECT}/user_data/runs")
-PATH_TO_WORKFLOWS = Path(f"{PATH_TO_PROJECT}/user_data/workflows")
-
-# color schemes
-PROTZILLA_DISCRETE_COLOR_SEQUENCE = [
-    "#4A536A",
-    "#87A8B9",
-    "#CE5A5A",
-    "#8E3325",
-    "#E2A46D",
-]
-PROTZILLA_DISCRETE_COLOR_OUTLIER_SEQUENCE = ["#4A536A", "#CE5A5A"]
-
-
-def random_string():
-    return "".join(choices(ascii_letters, k=16))
