+++ protzilla/data_preprocessing/plots.py
@@ -4,7 +4,7 @@ import plotly.graph_objects as go
 from plotly.graph_objects import Figure
 from plotly.subplots import make_subplots
 
-from ..constants.constants import (
+from ..constants.colors import (
     PROTZILLA_DISCRETE_COLOR_OUTLIER_SEQUENCE,
     PROTZILLA_DISCRETE_COLOR_SEQUENCE,
 )
